- 방문자 통계 오류 개선
- 장바구니 상품 수 실시간 조작
- 회원가입 쿠폰, 포인트 등록
- 관리자 주문 페이지
- 마이페이지 판매자 정보(모달)
- 마이페이지 문의하기(모달)

+++ IntelliJ/lotteOn/src/main/java/kr/co/lotteon/dto/delivery/DeliveryDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 public class DeliveryDTO {
 
-    private int dno;
+    private long dno;
 
     private int orderNo; // 주문번호
